refactor(service): CategoryTemplateService 삭제
