+++ backend/src/main/java/codezap/category/service/CategoryService.java
@@ -12,6 +12,7 @@ import codezap.category.dto.response.FindAllCategoriesResponse;
 import codezap.category.repository.CategoryRepository;
 import codezap.global.exception.CodeZapException;
 import codezap.member.domain.Member;
+import codezap.template.repository.TemplateRepository;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -19,13 +20,14 @@ import lombok.RequiredArgsConstructor;
 public class CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final TemplateRepository templateRepository;
 
     public CreateCategoryResponse create(Member member, CreateCategoryRequest createCategoryRequest) {
         String categoryName = createCategoryRequest.name();
         validateDuplicatedCategory(categoryName, member);
         Category category = categoryRepository.save(new Category(categoryName, member));
         return CreateCategoryResponse.from(category);
-}
+    }
 
     public FindAllCategoriesResponse findAllByMember(Member member) {
         return FindAllCategoriesResponse.from(categoryRepository.findAllByMemberOrderById(member));
@@ -50,6 +52,26 @@ public class CategoryService {
     private void validateDuplicatedCategory(String categoryName, Member member) {
         if (categoryRepository.existsByNameAndMember(categoryName, member)) {
             throw new CodeZapException(HttpStatus.CONFLICT, "이름이 " + categoryName + "인 카테고리가 이미 존재합니다.");
+        }
+    }
+
+    @Transactional
+    public void deleteById(Member member, Long id) {
+        Category category = categoryRepository.fetchById(id);
+        validateAuthorizedMember(category, member);
+
+        if (templateRepository.existsByCategoryId(id)) {
+            throw new CodeZapException(HttpStatus.BAD_REQUEST, "템플릿이 존재하는 카테고리는 삭제할 수 없습니다.");
+        }
+        if (category.getIsDefault()) {
+            throw new CodeZapException(HttpStatus.BAD_REQUEST, "기본 카테고리는 삭제할 수 없습니다.");
+        }
+        categoryRepository.deleteById(id);
+    }
+
+    private void validateAuthorizedMember(Category category, Member member) {
+        if (!member.equals(category.getMember())) {
+            throw new CodeZapException(HttpStatus.FORBIDDEN, "해당 카테고리를 수정 또는 삭제할 권한이 없는 유저입니다.");
         }
     }
 }
